Refine ambient heroes for core sections
